refactor to remove private util def
